+++ src/routes/home/home.component.jsx
@@ -1,0 +1,38 @@
+//import './categories.styles.scss';
+//import CategoryItem from './components/category-item/category-item.component';
+import Directory from "../../components/directory/directory.component";
+
+const Home=()=> {
+  const categories = [
+    {
+      "id": 1,
+      "title": "hats",
+      "imageUrl": "https://i.ibb.co/cvpntL1/hats.png"
+    },
+    {
+      "id": 2,
+      "title": "jackets",
+      "imageUrl": "https://i.ibb.co/px2tCc3/jackets.png"
+    },
+    {
+      "id": 3,
+      "title": "sneakers",
+      "imageUrl": "https://i.ibb.co/0jqHpnp/sneakers.png"
+    },
+    {
+      "id": 4,
+      "title": "women`s",
+      "imageUrl": "https://i.ibb.co/GCCdy8t/womens.png"
+    },
+    {
+      "id": 5,
+      "title": "men`s",
+      "imageUrl": "https://i.ibb.co/R70vBrQ/men.png"
+    }
+  ] ;
+  return (    
+    <Directory categories={categories}></Directory>
+  );
+}
+
+export default Home;
